account-service/requirements.txt
grpcio 
grpcio-tools
datetime
flask
waitress
python-dotenv